Improve MySQL query
Improve MySQL query by updating the `account_id` as well, on a duplicate
key constraint.
This ensures that, for example, if a new account signs in on the same device
the token created won't belong to the older account (assuming the older
account's token wasn't deleted from the server database).